feat: Implement MoviesActors' view and controller actions logic
Extend Index method in MoviesActorsController to accept actions
for fetching data from database.

Add ViewModel class with lists for movies and actors and action
description to pass it in MoviesActors view.

Refactor namespaces of model classes.

Add ActionLinks in index.html and show lists of movies
or actors depending on performed action.